split driver package into separate files
